Break dependency of ethcfg package to `txpool`. Create `txpoolcfg` package (#958)
